Decompose backend hook file
