+++ scripts/file-length-allowlist.txt
@@ -17,4 +17,3 @@
 # be split, not grandfathered.
 
 # --- production code (limit 1000) ---
-session/backend_hook.go 1026
